Stop using VLA with ancillary data
Using VLA is not good since tcp_connection doesn't have any hard limit
on the amount of ancillary data that can be sent with it.
Reserving memory for this data with VLA may cause the application to
crash if it needs to allocate more memory than VLA allows.

This commit changes this usage to use dynamic allocation which should
avoid this problem.