tcp_connection: do not defer closing of the output stream
At present, tcp_close and tcp_set_output_stream defer the close of the
output stream. This allows unsafe call chains such as the ones in the
tcp_client object, but makes it hard to manage the lifetime of an
object on top of tcp_conn_t, since it is not possible to assume that,
after tcp_close is called, no callback with a reference to the object
will be scheduled.

This commit changes tcp_close and tcp_set_output_stream to perform an
immediate close of the output stream. While the change is risky, code
that relies on the current behaviour can (and should) be rewritten so
as to not assume any specific close semantics.