jsondecoder: fix reading on EOF
jsondecoder does not correctly detect EOF from the underlying stream.
This causes an assertion failure if the application calls
jsondecoder_receive again after having received a JSON object or an
EILSEQ error. To fix the issue, this commit changes jsondecoder to use
a byte array for buffering, which makes it easier to detect the EOF.