jsonserver: fix crash introduced in 03f402f
The farewell stream wrapping a connection output queuestream is now
relaxed. Because of this, it is possible for the connection input
callback to be executed after the write side of the connection is
closed but before the farewell stream callback is executed. In this
case, where the connection is still in open state but its output
queuestream is closed, if the read side of the connection is also
closed, the input callback tries to terminate the already closed
output queuestream, which causes an assertion failure as the
queuestream is not relaxed.