test: Use tlog_string() for printing empty lines
asynctest-fsadns.c intentionally wants to print empty test log line to
make the output easier to read. However, using tlog with empty string
gives a warning from GCC about empty format strings.

Several options to resolve those were considered:
- Adding repo wide -Wno-format-zero-length compiler flag
- Disabling this warning explicitly in those few places or in the one
  source file with:
  ```
  #pragma GCC diagnostics push
  #pragma GCC diagnostics ignored "-Wformat-zero-length"
  ...
  #pragma GCC diagnostics pop
  ```
- Using non-empty whitespace-only format string like `tlog(" ");`
- Adding a separate log function for this use case

The last option was implemented in this commit as it's the most clear
solution and it doesn't limit the fix to single source file. And it
doesn't remove that warning detection for the whole repo.