test: do not call fork in a multithreaded process
The async_loop_protected test makes the process multithreaded. This
may cause subsequent tests of async modules that use fork() to
deadlock or crash, because fork() is executed in a multithreaded
environment. This commit introduces a separate list of tests that
spawn threads and changes asynctest to run these tests _after_ all the
single-threaded tests.